task1 %c and %s

+++ main.h
@@ -3,6 +3,9 @@
 
 #include <stdarg.h>
 #include <unistd.h>
+#include <limits.h>
+#include <stdio.h>
+
 
 int _printf(const char *format, ...);
 
@@ -11,6 +14,6 @@ int	ft_putnbr(long int n);
 int	ft_putstr(char *str);
 int	ft_unsignedint(unsigned int n);
 int	ft_hexadecimal(unsigned int n, char *base);
-int	ft_adresse(long long unsigned int n, char *base);
+int	ft_adresse(long unsigned int n, char *base);
 
 #endif
